# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PingResult {
  /** Whether the connection to ServiceNow succeeded */
  ok: boolean;
  /** The ServiceNow instance URL that was tested */
  instance: string;
  /** Round-trip latency in milliseconds */
  latencyMs: number;
  /** Error message if ok=false */
  error?: string;
}

export interface ServicenowConfig {
  /** The configured ServiceNow instance URL */
  instance: string;
}

export interface ApiError {
  error: string;
}

export interface ReportConfig {
  id: number;
  name: string;
  sysId: string;
  filterQuery: string;
  fields: string;
  verifiedTitle?: string | null;
  verifiedTable?: string | null;
  verifiedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateReportRequest {
  name: string;
  sysId: string;
  filterQuery?: string;
  fields?: string;
}

export interface UpdateReportRequest {
  name?: string;
  sysId?: string;
  filterQuery?: string;
  fields?: string;
}