# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: servicenow
    description: ServiceNow connectivity operations
  - name: reports
    description: Report configuration management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /servicenow/config:
    get:
      operationId: servicenowConfig
      tags: [servicenow]
      summary: Get ServiceNow configuration
      description: Returns the configured ServiceNow instance URL (non-sensitive)
      responses:
        "200":
          description: Configuration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicenowConfig"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /servicenow/ping:
    get:
      operationId: servicenowPing
      tags: [servicenow]
      summary: Test ServiceNow connectivity
      description: Makes a lightweight request to the configured ServiceNow instance and returns connection status
      responses:
        "200":
          description: Connection result (ok=true means connected, ok=false means failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PingResult"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all report configurations
      responses:
        "200":
          description: List of report configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportConfig"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a new report configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportRequest"
      responses:
        "201":
          description: Created report config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportConfig"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/{id}:
    put:
      operationId: updateReport
      tags: [reports]
      summary: Update an existing report configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportRequest"
      responses:
        "200":
          description: Updated report config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportConfig"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/{id}/verify:
    post:
      operationId: verifyReport
      tags: [reports]
      summary: Verify a report config against ServiceNow
      description: Fetches the report definition from sys_report table and stores the resolved title and table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportConfig"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PingResult:
      type: object
      properties:
        ok:
          type: boolean
          description: Whether the connection to ServiceNow succeeded
        instance:
          type: string
          description: The ServiceNow instance URL that was tested
        latencyMs:
          type: number
          description: Round-trip latency in milliseconds
        error:
          type: string
          description: Error message if ok=false
      required:
        - ok
        - instance
        - latencyMs
    ServicenowConfig:
      type: object
      properties:
        instance:
          type: string
          description: The configured ServiceNow instance URL
      required:
        - instance
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ReportConfig:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sysId:
          type: string
        filterQuery:
          type: string
        fields:
          type: string
        verifiedTitle:
          type: string
          nullable: true
        verifiedTable:
          type: string
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sysId
        - filterQuery
        - fields
        - createdAt
        - updatedAt
    CreateReportRequest:
      type: object
      properties:
        name:
          type: string
        sysId:
          type: string
        filterQuery:
          type: string
          default: ""
        fields:
          type: string
          default: ""
      required:
        - name
        - sysId
    UpdateReportRequest:
      type: object
      properties:
        name:
          type: string
        sysId:
          type: string
        filterQuery:
          type: string
        fields:
          type: string