# lib/api-zod/src/generated/types/createReportRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReportRequest {
  name: string;
  sysId: string;
  filterQuery?: string;
  fields?: string;
}