# lib/api-zod/src/generated/types/reportConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportConfig {
  id: number;
  name: string;
  sysId: string;
  filterQuery: string;
  fields: string;
  verifiedTitle?: string | null;
  verifiedTable?: string | null;
  verifiedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}